Revert "FC Diff view (all fields) [2/2] (#2678)"
This reverts commit 467181a28206aee624d56f56b4ffb8d6b2ac4ed7.
