--- src/components/v2v3/V2V3Project/V2V3FundingCycleSection/FundingCycleDetails/RulesListItems/PausePayValue.tsx
@@ -1,5 +1,0 @@
-import { Trans } from '@lingui/macro'
-
-export function PausePayValue({ pausePay }: { pausePay: boolean }) {
-  return pausePay ? <Trans>Paused</Trans> : <Trans>Enabled</Trans>
-}
