+++ src/components/v2v3/V2V3Project/V2V3FundingCycleSection/FundingCycleDetails/index.tsx
@@ -45,7 +45,6 @@ export default function FundingCycleDetails({
           <TokenListItems
             fundingCycle={fundingCycle}
             fundingCycleMetadata={fundingCycleMetadata}
-            showDiffs={showDiffs}
           />
         }
       />
@@ -55,7 +54,6 @@ export default function FundingCycleDetails({
           <RulesListItems
             fundingCycle={fundingCycle}
             fundingCycleMetadata={fundingCycleMetadata}
-            showDiffs={showDiffs}
           />
         }
       />
